requirements.txt
nixtla
pandas
python-dotenv
matplotlib
